Add type hints to generator
