+++ skale/types/node.py
@@ -17,8 +17,43 @@
 #   You should have received a copy of the GNU Affero General Public License
 #   along with SKALE.py.  If not, see <https://www.gnu.org/licenses/>.
 
-from typing import NewType
+from enum import IntEnum
+from typing import List, NewType, TypedDict
+
+from eth_typing import BlockNumber
+
+from skale.types.schain import SchainStructureWithStatus
+from skale.types.validator import ValidatorId
 
 
 NodeId = NewType("NodeId", int)
 Port = NewType("Port", int)
+
+
+class NodeStatus(IntEnum):
+    ACTIVE = 0
+    LEAVING = 1
+    LEFT = 2
+    IN_MAINTENANCE = 3
+
+
+class Node(TypedDict):
+    name: str
+    ip: bytes
+    publicIP: bytes
+    port: Port
+    start_block: BlockNumber
+    last_reward_date: int
+    finish_time: int
+    status: NodeStatus
+    validator_id: ValidatorId
+    publicKey: str
+    domain_name: str
+
+
+class NodeWithId(Node):
+    id: NodeId
+
+
+class NodeWithSchains(NodeWithId):
+    schains: List[SchainStructureWithStatus]
